Assigned roles to users

+++ app/Http/Controllers/Admin/RoleController.php
@@ -37,7 +37,7 @@ class RoleController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
@@ -46,7 +46,14 @@ class RoleController extends Controller
         // dd($request['name'])
 
         $data = $request->all();
-        // dd($data);
+        // dd($data['name']);
+
+        // Verifica se o nome é unico
+        if ((boolean)Role::where('name', '=', $data['name'])->first()) {
+            $messageError = 'Já existe um papel com o nome: ' . $data['name'];
+            return $messageError;
+        }
+
         Role::create($data);
 
         return redirect()->route('admin.roles'); // !Não precisa das vars $item e $goToSection, a rota é chamada
@@ -55,7 +62,7 @@ class RoleController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Role  $role
+     * @param  \App\Role $role
      * @return \Illuminate\Http\Response
      */
     public function show(Role $role, $id)
@@ -71,7 +78,7 @@ class RoleController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Role  $role
+     * @param  \App\Role $role
      * @return \Illuminate\Http\Response
      */
     public function edit(Role $role, $id)
@@ -85,14 +92,28 @@ class RoleController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Role  $role
+     * @param  \Illuminate\Http\Request $request
+     * @param  \App\Role $role
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Role $role, $id)
     {
         $data = $request->all();
         // dd($data);
+
+        // Verifica se o nome não é igual ao de outro papel
+        $roleCheck = Role::where('name', '=', $data['name'])->first();
+
+        if (isset($roleCheck) && ($roleCheck->id != $id)) {
+            $messageError = 'Já existe um papel com o nome: ' . $data['name'];
+            return $messageError;
+        }
+
+        // Verifica se não é o SuperAdmin
+        if ($id == 1) {
+            $messageError = 'Este papel não pode ser alterado!';
+            return $messageError;
+        }
 
         // dd($curso->id); // Por segurança buscar pelo id originário($curso) e não o enviado($request)
         Role::find($id)->update($data);
@@ -104,11 +125,17 @@ class RoleController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Role  $role
+     * @param  \App\Role $role
      * @return \Illuminate\Http\Response
      */
     public function destroy(Role $role, $id)
     {
+        // Verifica se não é o SuperAdmin
+        if ($id == 1) {
+            $messageError = 'Este papel não pode ser apagado!';
+            return $messageError;
+        }
+
         Role::find($id)->delete();
         return redirect()->route('admin.roles');
     }
